sc_depth/requirements.txt
pytorch-lightning==1.7.3
matplotlib
opencv-python
tqdm
imageio
path
scipy
configargparse

# conda create -n sc_depth_env python=3.8
# conda activate sc_depth_env
# conda install pytorch torchvision pytorch-cuda=11.7 -c pytorch -c nvidia
# pip install -r requirements.txt